backend/python/requirements.txt
flask
pytesseract
Pillow
transformers
torch
